requirements.txt
Flask==2.1.2
Flask-RESTful==0.3.9
python-dotenv==0.20.0
gunicorn==20.1.0
Werkzeug==2.2.3
openai==1.52.2
